+++ src/main/java/com/ebicep/warlords/player/ingame/WarlordsNPC.java
@@ -274,7 +274,7 @@ public final class WarlordsNPC extends WarlordsEntity {
     }
 
     @Override
-    public void setDamageResistance(int damageResistance) {
+    public void setDamageResistance(float damageResistance) {
         getSpec().setDamageResistance(Math.max(0, damageResistance));
         updateHealth();
 //        nameDisplay.customName(getNameComponent());
